#### What type of PR is this?
optimize

#### What this PR does / why we need it (English/Chinese):
zh: 减少 grpc 部分的 ut 的时长，从 97s 减少到 35s
en: reduce the grpc uts timecost from 97s to 35s

#### Which issue(s) this PR fixes: